+++ disk_cache_manager/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+
+- Add exists method.
+
 ## 1.0.1
 
 - Update dependencies.